Mystoreproject/src/main/java/com/mystore/pageObjects/AccountCreationPage.java
package com.mystore.pageObjects;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.mystore.ActionDrivers.Action;
import com.mystore.base.BaseClass;

public class AccountCreationPage extends BaseClass {

	@FindBy(xpath="//h1[contains(text(),'Create an account')]")
	WebElement formtitle;
	
	public AccountCreationPage() {
		PageFactory.initElements(getDriver(), this);
	}
	
	public boolean checkformtitle() {
		return Action.isDisplayed(getDriver(), formtitle);
	}
}
